subtitle_translator/requirements.txt
PyQt5>=5.15.0
Pillow>=9.0.0
pytesseract>=0.3.10
tencentcloud-sdk-python>=3.0.0
opencv-python>=4.5.0
numpy>=1.21.0
